chore(middleware): move two classes into one file
